requirements.txt
pyyaml
tqdm
soundfile
librosa == 0.8.1
